requirements.txt
openai>=1.0.0      # 调用大模型API的基础库
python-dotenv      # 用于管理API Key等环境变量
streamlit          # 用于快速搭建前端交互界面
pydantic           # 用于规范模型输出的JSON格式